Move symbol extension functions into separate `SymbolExt`
